completed jwt code and implemented delete of post using token

+++ controller/api/v1/posts_api.js
@@ -25,17 +25,23 @@ module.exports.destroy=async function(req,res){
     try{
       let post=await Post.findById(req.params.id);
       // .id converting the object id into string
+       if(post.user==req.user.id){
           post.remove();
           await Comment.deleteMany({post: req.params.id});
           
           return res.json(200,{
               message: "Post and associated comment deleted successfully"
           });
+       }else{
+           return res.json(401,{
+               message:"You cannot delete this post"
+           })
+       }   
+       }
+    catch(err){
+        console.log(err);
+        return res.json(500,{
+            message: "Internal server error"
+        }); 
     }
-          catch(err){
-          console.log(err);
-          return res.json(500,{
-              message: "Internal server error"
-          });
-        }
 }
